Rename extractor to archiver
